+++ src/lib/tests/book.rs
@@ -11,7 +11,7 @@ description: >-
  long
  description
 epub.version: 3";
-    let mut book = Book::new();
+    let mut book = Book::new(&[]);
     book.set_from_config(config).unwrap();
     test_eq(book.options.get_str("author").unwrap(), "Author");
     test_eq(book.options.get_str("title").unwrap(), "Some title");